Log module creation. This will help with "weird crashes".
A lot of crashes related to WhatTheRack are due to modules crashing
when they're constructed. This appears to be WTR's fault as the user
clicks the button and Rack terminates with no explanation. Logs will
help pinpoint culprit modules. I may start a blacklist of sorts, to
filter them out.